+++ entity/src/main/java/io/github/fabricators_of_create/porting_lib/entity/events/EntityTeleportCallback.java
@@ -1,6 +1,6 @@
 package io.github.fabricators_of_create.porting_lib.entity.events;
 
-import io.github.fabricators_of_create.porting_lib.core.event.CancellableEvent;
+import io.github.fabricators_of_create.porting_lib.core.event.object.CancellableEvent;
 import net.fabricmc.fabric.api.event.Event;
 import net.fabricmc.fabric.api.event.EventFactory;
 import net.minecraft.world.entity.Entity;
@@ -30,7 +30,7 @@ public interface EntityTeleportCallback {
 	 */
 	void onTeleport(EntityTeleportEvent event);
 
-	class EntityTeleportEvent extends CancellableEvent.Base {
+	class EntityTeleportEvent extends CancellableEvent {
 		public final Entity entity;
 		public double targetX, targetY, targetZ;
